chore(ts): init typescript conf + tsify tokenize-arg-string
