+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
+build/
 .nyc_output
 node_modules
 .DS_Store
