+++ .mocharc.json
@@ -1,0 +1,6 @@
+{
+  "spec": [
+    "build/test",
+    "test"
+  ]
+}
