+++ .nycrc
@@ -1,0 +1,13 @@
+{
+  "exclude": [
+    "build/test/**",
+    "test/**"
+  ],
+  "reporter": [
+    "html",
+    "text"
+  ],
+  "lines": 100,
+  "branches": "97",
+  "statements": "100"
+}
